Create AlgebraTutorV1
This is a three part staged assignment.  You must complete part 1 with at least 80% success  before you can begin part 2, and the same with part 3.  After you've completed and submitted each part, you can proceed to the next part by selecting the drop down arrow next to the assignment title. You will need to complete all parts to receive full credit.

Part 1. Solve for y.  Write a program with a that displays a randomly generated problem that asks the user to solve for the y variable, takes input from the user, and prints if the user answered correctly or not.   Your main should give one problem and then exit.  Use one or more methods to produce this behavior.

Example interactions:

Given:

m = 10
x = 3
b = 2

What is the value of y? 32
Correct!

Given:

m = 6
x = 9
b = 3

What is the value of y? 52
Sorry, that is incorrect. The answer is 57.

Part 2. Solve for y, m and b.  Add 2 choices that behave similarly to your part 1 program, but have the user solve for m and b instead.  These choices should also randomly generate problems, and print if the user answers correctly.  Your main program should display a menu asking which type of question to ask:  Select 1 to Solve for Y, 2 to Solve for M, 3 to Solve for B and 4 to quit.  When the user selects the question type, one question should be asked, then the user should return to the menu.  This should repeat until the user selects quit, then the program should exit.  Use one or more methods and loops to produce this behavior.

Part 3. Three in a Row with Hints.  Update your program so that each problem type mode will repeatedly ask questions until the user gets 3 correct answers in a row.  If the attempts more than 3 questions in a particular mode, the program should provide a hint on how to solve problems of this type.   After the student has correctly answered 3 questions in a row, an overall score (the number of questions answered correctly divided by the total number of questions attempted) should be displayed, and the menu is presented again.